fix: SDL2 framework include on macOS
The build was broken on recent versions of the official SDL2
framework due to changes in the header layout. It still works fine
if using the recommended way to access framework headers, but not
when using -I to index into the framework.

To #include from the SDL2 framework, the #include statement needs
to start with the name of the framework and the appropriate
`-F` flags to the framework's location has to be specified
instead of `-I`.